bug review

+++ scripts/shell-based.sh
@@ -18,6 +18,7 @@ curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh | 
 echo 'eval "$(/home/linuxbrew/.linuxbrew/bin/brew shellenv)"' | sudo -u vagrant tee -a /home/vagrant/.zprofile
 
 # allow setting of DISPLAY variable for Windows XServer
+# TODO this should be moved to ANSIBLE to make it idempotent
 echo export DISPLAY='$(/sbin/ip route | awk '"'"'/default/ { print $3 }'"'"'):0' | sudo -u vagrant tee -a /home/vagrant/.zprofile 
 
 # McFly (advanced shell history)
@@ -31,7 +32,6 @@ pip3 install ansible ansible-lint yq tmuxp tig pre-commit
 
 # install docker
 # see https://docs.docker.com/engine/install/ubuntu/
-# TODO move to ansible
 echo "Install Docker..."
 curl -fsSL https://download.docker.com/linux/ubuntu/gpg | apt-key add -
 add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu focal stable"
@@ -44,11 +44,11 @@ usermod -aG docker vagrant
 echo "Install Hashicorp Tools..."
 curl -fsSL https://apt.releases.hashicorp.com/gpg | sudo apt-key add -
 apt-add-repository "deb [arch=amd64] https://apt.releases.hashicorp.com $(lsb_release -cs) main"
-apt-get update && sudo apt-get install -y consul nomad vault terraform consul-template
-sudo -u vagrant zsh -c boundary config autocomplete install
-sudo -u vagrant zsh -c consul -autocomplete-install
-sudo -u vagrant zsh -c nomad -autocomplete-install
-sudo -u vagrant zsh -c terraform -install-autocomplete
+apt-get update && sudo apt-get install -y consul nomad vault terraform consul-template boundary
+sudo -H -u vagrant zsh -c 'cd /home/vagrant ; boundary config autocomplete install'
+sudo -H -u vagrant zsh -c 'cd /home/vagrant ; consul -autocomplete-install'
+sudo -H -u vagrant zsh -c 'cd /home/vagrant ; nomad -autocomplete-install'
+sudo -H -u vagrant zsh -c 'cd /home/vagrant ; terraform -install-autocomplete'
 
 # AWS CLI
 curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
@@ -126,12 +126,6 @@ sudo -u vagrant sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/maste
 echo "Install sdkman..."
 sudo -u vagrant sh -c "curl -s https://get.sdkman.io | bash"
 
-# install nvm (managing node versions)
-echo "Install nvm..."
-#sudo -u vagrant sh -c "curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.37.2/install.sh | bash"
-git clone https://github.com/lukechilds/zsh-nvm ~/.oh-my-zsh/custom/plugins/zsh-nvm
-
-
 ## change shell to zsh
 echo "change vagrant shell to zsh..."
 chsh --shell /usr/bin/zsh vagrant
@@ -148,21 +142,14 @@ fi
 ## brew installs
 echo "installing multiple brews..."
 sudo -H -i -u vagrant zsh -c "brew install lsd gitui lazygit git-delta procs broot rs/tap/curlie derailed/k9s/k9s terragrunt cdktf"
+
 # set alias
+# TODO this should be moved to ansible to make it idempotent
 echo "alias ls='lsd'" >> .oh-my-zsh/custom/alias.zsh
+echo "alias bat='batcat'" >> .oh-my-zsh/custom/alias.zsh
 # kubeconfig for k3s
 sudo -u vagrant mkdir /home/vagrant/.kube
 sudo -u vagrant ln -s /etc/rancher/k3s/k3s.yaml /home/vagrant/.kube/config
-
-# Git
-# TODO move to Ansible
-echo "configure git"
-sudo -u vagrant git config --global core.pager delta
-sudo -u vagrant git config --global interactive.diffFilter "delta --color-only"
-sudo -u vagrant git config --global credential.helper store
-sudo -u vagrant git config --global core.editor vim
-sudo -u vagrant git config --global commit.template ~/.gitmessage
-sudo -u vagrant cp ansible/playbooks/files/git-commit-template.txt /home/vagrant/.gitmessage
 
 echo "install nerd font..."
 bash -s nerd-font.sh
@@ -171,19 +158,22 @@ bash -s nerd-font.sh
 # exporting vagrant home and /etc/rancher as shares.
 # the latter can be used to add config to lens as a kube config
 # need to mount shares and assign a device letter.
+# TODO move to Ansible
 echo "install and configure samba"
 apt-get install -y samba
-cp files/smb.conf /etc/samba/smb.conf
+cp /vagrant/files/smb.conf /etc/samba/smb.conf
 systemctl restart smbd
 # create smb user vagrant/vagrant
 (echo vagrant; echo vagrant) | sudo smbpasswd -a -s vagrant
 
 # Lens
 echo "install lens"
-LENS_FILE=Lens-5.3.0-latest.20211125.2.amd64.deb
-wget "https://api.k8slens.dev/binaries/${LENS_FILE}"
-apt install -y libnotify4 libnss3 libxss1 xdg-utils libsecret-1-0 libnspr4 libsecret-common libxshmfence1 libgbm1  
-dpkg -i "${LENS_FILE}"
+if [ ! -f /usr/bin/lens ] ; then
+  LENS_FILE=Lens-5.3.0-latest.20211125.2.amd64.deb
+  wget -q "https://api.k8slens.dev/binaries/${LENS_FILE}"
+  apt install -y libnotify4 libnss3 libxss1 xdg-utils libsecret-1-0 libnspr4 libsecret-common libxshmfence1 libgbm1  
+  dpkg -i "${LENS_FILE}"
+fi
 
 # clone
 echo "cloning repo..."
@@ -197,6 +187,5 @@ fi
 
 # Run Ansible
 # Workaround youcomplete me
-ln -s /bin/gcc /bin/gcc-5
 echo "running ansible"
-sudo -u vagrant sh -c cd git/l4win/ansible && ansible-playbook -i inventory.yml playbooks/site.yml
+sudo -H -u vagrant sh -c 'unset SUDO_USER; unset SUDO_COMMAND; env ; cd /home/vagrant/.vim/bundle/YouCompleteMe ; ./install.py'
